// frontend/src/api/generated/models/siteIncidentResolutionEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Dekopen API
 * Authenticated tenant and engine API boundary.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `REMAKE` - REMAKE
 * * `PURCHASE` - PURCHASE
 * * `SERVICE` - SERVICE
 * * `NONE` - NONE
 */
export type SiteIncidentResolutionEnum =
  (typeof SiteIncidentResolutionEnum)[keyof typeof SiteIncidentResolutionEnum];

export const SiteIncidentResolutionEnum = {
  REMAKE: "REMAKE",
  PURCHASE: "PURCHASE",
  SERVICE: "SERVICE",
  NONE: "NONE",
} as const;